requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
environs==9.5.0
marshmallow==3.20.1
packaging==23.1
psycopg2==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
gunicorn
drf-yasg
django-filter
django-silk
